Added try-finally in lifespan app.py
